feat(shopping-cart): integrate cart-service with frontend for add/update/remove cart items

+++ cart-service/src/main/java/com/tqk/cartservice/dto/request/AddCartItemRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class AddCartItemRequest {
-    private Long productId;
-    private Long variantId;
+    private Integer productId;
+    private Integer variantId;
     private Integer quantity;
 }
